Draft: Added hex dump and tracing of send/recv buffers
Down in kmip_bio_send_request_encoding is where the meat of the the
sending and receiving buffers to/from the KMIP server occurs for all the
low-level APIs. While debugging an interoperability issue with a
specific KMIP vendor's product, it was discovered that the server was
terminating the connection early, truncating the response by some number
of bytes from what it communicated in the length. This patch adds hex
dumping of the BIO_read and BIO_write calls, so that we can do forensics
on the raw buffer and try to figure out any discrepancies or
encoding/decoding bugs on either side of the connection.

This is a draft. A future commit will make this functionality able to be
turned on dynamically, but off by default.